+++ src/test/java/com/somemore/domains/volunteer/service/UpdateVolunteerLockServiceTest.java
@@ -13,7 +13,7 @@ import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
-import static com.somemore.global.auth.oauth.OAuthProvider.NAVER;
+import static com.somemore.global.auth.oauth.domain.OAuthProvider.NAVER;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class UpdateVolunteerLockServiceTest extends IntegrationTestSupport {
