+++ src/test/java/com/somemore/domains/volunteer/service/VolunteerQueryServiceTest.java
@@ -8,7 +8,7 @@ import com.somemore.domains.volunteer.dto.response.VolunteerRankingResponseDto;
 import com.somemore.domains.volunteer.repository.VolunteerDetailRepository;
 import com.somemore.domains.volunteer.repository.VolunteerRepository;
 import com.somemore.domains.volunteer.repository.mapper.VolunteerSimpleInfo;
-import com.somemore.global.auth.oauth.OAuthProvider;
+import com.somemore.global.auth.oauth.domain.OAuthProvider;
 import com.somemore.global.exception.BadRequestException;
 import com.somemore.support.IntegrationTestSupport;
 import org.junit.jupiter.api.DisplayName;
